move tests out of URLNavigator tests
